htgen/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   google-genai
    #   httpx
astroid==3.3.10
    # via pylint
black==25.1.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   black
    #   flask
    #   pip-tools
dill==0.4.0
    # via pylint
exif==1.6.1
    # via -r requirements.in
flask==3.1.1
    # via -r requirements.in
google-api-core==2.25.1
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
    #   google-genai
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.1
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.23.0
    # via -r requirements.in
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via google-genai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
isort==6.0.1
    # via pylint
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via pylint
mypy-extensions==1.1.0
    # via black
packaging==25.0
    # via
    #   black
    #   build
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.8
    # via
    #   black
    #   pylint
plum-py==0.8.7
    # via exif
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.7
    # via google-genai
pydantic-core==2.33.2
    # via pydantic
pylint==3.3.7
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via -r requirements.in
python-magic==0.4.27
    # via -r requirements.in
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==8.5.0
    # via google-genai
tomlkit==0.13.3
    # via pylint
typing-extensions==4.14.0
    # via
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via google-genai
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
